rag_chat_local/requirements.txt
chromadb
llama-cpp-python
gradio
pyyaml
azure-search-documents
pypdf